requirements.txt
requests
json
pyyaml
openai
praw